% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnc.R
\name{cnc_pags}
\alias{cnc_pags}
\title{Baixa páginas do CNC}
\usage{
cnc_pags(path = "data-raw/pags", pags = 1:2399)
}
\arguments{
\item{path}{caminho da pasta onde os arquivos HTML serão salvos. Se a pasta não existir, será criada.}

\item{pags}{páginas a serem baixadas. Default 1:2399 (não atualiza com muita frequênica)}
}
\value{
\code{data.frame} indicando "OK" se baixou corretamente e "arquivo existe" se o arquivo já existe.
}
\description{
Acessa todas as páginas do CNC indicadas e salva em arquivos HTML. Acessa páginas do tipo
\url{http://www.cnj.jus.br/improbidade_adm/consultar_requerido.php?validar=form&rs=pesquisarRequeridoGetTabela&rst=&rsrnd=0&rsargs[]=&rsargs[]=&rsargs[]=&rsargs[]=&rsargs[]=&rsargs[]=&rsargs[]=I&rsargs[]=0&rsargs[]=POSICAO_INICIAL_PAGINACAO_PHP1&rsargs[]=QUANTIDADE_REGISTROS_PAGINACAO15}
}
