% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnc.R
\name{cnc_pessoas}
\alias{cnc_pessoas}
\title{Baixa pessoas}
\usage{
cnc_pessoas(dados, path = "data-raw/pessoas")
}
\arguments{
\item{dados}{retornados pela função \code{\link{parse_pags}}.}

\item{path}{caminho da pasta onde os arquivos HTML serão salvos. Se a pasta não existir, será criada.}
}
\value{
\code{data.frame} indicando "OK" se baixou corretamente e "arquivo existe" se o arquivo já existe.
}
\description{
Baixa HTMLs de pesquisas do tipo
\url{http://www.cnj.jus.br/improbidade_adm/visualizar_condenacao.php?seq_condenacao=1101}
}

