% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnc.R
\name{cnc_pessoas_infos}
\alias{cnc_pessoas_infos}
\title{Baixa infos de pessoas}
\usage{
cnc_pessoas_infos(d_pessoas, path = "data-raw/pessoas_infos")
}
\arguments{
\item{d_pessoas}{retornados pela função \code{\link{parse_pessoas}}.}

\item{path}{caminho da pasta onde os arquivos HTML serão salvos. Se a pasta não existir, será criada.}
}
\value{
\code{data.frame} indicando "OK" se baixou corretamente e "arquivo existe" se o arquivo já existe.
}
\description{
Baixa HTMLs de pesquisas do tipo
\url{http://www.cnj.jus.br/improbidade_adm/visualizar_condenacao.php?seq_condenacao=1&rs=getDadosParte&rst=&rsrnd=0&rsargs[]=1}
}
