% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnc.R
\name{cnc_processos}
\alias{cnc_processos}
\title{Baixa processos}
\usage{
cnc_processos(dados, path = "data-raw/processos")
}
\arguments{
\item{dados}{retornados pela função \code{\link{parse_pags}}.}

\item{path}{caminho da pasta onde os arquivos HTML serão salvos. Se a pasta não existir, será criada.}
}
\value{
\code{data.frame} indicando "OK" se baixou corretamente e "arquivo existe" se o arquivo já existe.
}
\description{
Baixa HTMLs de pesquisas do tipo
\url{http://www.cnj.jus.br/improbidade_adm/visualizar_processo.php?seq_processo=9421}
}
