% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{cnc_npag}
\alias{cnc_npag}
\alias{download}
\alias{cnc_download_pag}
\alias{cnc_download_pessoa}
\alias{cnc_download_processo}
\alias{cnc_download_pessoa_infos}
\title{Download CNC}
\usage{
cnc_npag()

cnc_download_pag(pag, path)

cnc_download_pessoa(link, path)

cnc_download_processo(link, path)

cnc_download_pessoa_infos(link, path)
}
\arguments{
\item{pag}{página a ser baixada.}

\item{path}{caminho da pasta onde os arquivos HTML serão salvos. Se a pasta não existir, será criada.}

\item{link}{retornado pela função \code{\link[=cnc_parse_pag]{cnc_parse_pag()}}.}
}
\value{
\code{\link[=cnc_npag]{cnc_npag()}} retorna o número de páginas que podem ser
consultadas no CNC.

As demais funções retornam o caminho do arquivo baixado.
}
\description{
\code{\link[=cnc_npag]{cnc_npag()}} acessa a pagina do CNC com uma consulta genérica e coleta o
número de páginas disponível para consulta.

\code{\link[=cnc_download_pag]{cnc_download_pag()}} acessa uma página do CNC indicada e salva em arquivos HTML.
Acessa páginas do tipo
\url{http://www.cnj.jus.br/improbidade_adm/consultar_requerido.php?validar=form&rs=pesquisarRequeridoGetTabela&rst=&rsrnd=0&rsargs[]=&rsargs[]=&rsargs[]=&rsargs[]=&rsargs[]=&rsargs[]=&rsargs[]=I&rsargs[]=0&rsargs[]=POSICAO_INICIAL_PAGINACAO_PHP1&rsargs[]=QUANTIDADE_REGISTROS_PAGINACAO15}

\code{\link[=cnc_download_pessoa]{cnc_download_pessoa()}} baixa HTML de pesquisas do tipo
\url{http://www.cnj.jus.br/improbidade_adm/visualizar_condenacao.php?seq_condenacao=1101}

\code{\link[=cnc_download_processo]{cnc_download_processo()}} baixa HTML de pesquisas do tipo
\url{http://www.cnj.jus.br/improbidade_adm/visualizar_processo.php?seq_processo=9421}

\code{\link[=cnc_download_pessoa_infos]{cnc_download_pessoa_infos()}} baixa HTML de pesquisas do tipo
\url{http://www.cnj.jus.br/improbidade_adm/visualizar_condenacao.php?seq_condenacao=1&rs=getDadosParte&rst=&rsrnd=0&rsargs[]=1}
}
