% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dvec}
\alias{dvec}
\title{Vetorizando parsers}
\usage{
dvec(fun, arqs, verbose = TRUE)
}
\arguments{
\item{fun}{função a ser aplicada em cada arquivo.}

\item{arqs}{character vector dos caminhos de arquivos a serem transformados.}

\item{verbose}{se \code{TRUE} (default), mostra o arquivo que está sendo lido com probabilidade de 5\%.}
}
\description{
Vetoriza um parser (função) para um vetor de arquivos.
}

