% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnc.R
\name{GET_pjs}
\alias{GET_pjs}
\title{Baixa uma página web qualquer usando PhantomJS}
\usage{
GET_pjs(url, arq = "arq.html")
}
\arguments{
\item{url}{endereço web que se deseja baixar.}

\item{arq}{caminho do arquivo para baixar.}
}
\value{
retorna \code{TRUE} se der tudo certo e ele conseguir criar e deletar todos os arquivos temporários.
}
\description{
Acessa uma URL, roda javascripts e baixa página usando PhantomJS.
Para funcionar, é necessário instalar o PhamtomJS
(\url{http://phantomjs.org/download.html}).
}

