% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse}
\alias{parse}
\alias{cnc_parse_pag}
\alias{cnc_parse_pessoa}
\alias{cnc_parse_processo}
\alias{cnc_parse_pessoa_infos}
\title{Parse dos arquivos HTML baixados}
\usage{
cnc_parse_pag(arq)

cnc_parse_pessoa(arq)

cnc_parse_processo(arq)

cnc_parse_pessoa_infos(arq)
}
\arguments{
\item{arq}{caminho do arquivo que deve ser processado.}
}
\value{
Para \code{\link[=cnc_parse_pag]{cnc_parse_pag()}}, uma \code{tibble} com as colunas

\itemize{
\item \code{id} id (1 a 15) do indivíduo obtido na página.
\item \code{key} \code{"nm_pessoa"} ou \code{"num_processo"}, indicando se é o nome da pessoa ou número do processo.
\item \code{value} nome da pessoa ou número do processo.
\item \code{link} para acessar informações da pessoa ou do processo (info utilizada nos outros scrapers).
}

Para \code{\link[=cnc_parse_pessoa]{cnc_parse_pessoa()}}, uma \code{tibble} com as colunas
\itemize{
\item \code{key} nome (label) da informação. Geralmente contém localização do processo, data da propositura e número do processo.
\item \code{value} conteúdo da informação.
}

Para \code{\link[=cnc_parse_processo]{cnc_parse_processo()}}, uma \code{tibble} com as colunas
\itemize{
\item \code{key} nome (label) da informação. Geralmente contém localização do processo, data da propositura e número do processo.
\item \code{value} conteúdo da informação.
}

Para \code{\link[=cnc_parse_pessoa_infos]{cnc_parse_pessoa_infos()}}, uma \code{tibble} com as colunas
\itemize{
\item \code{key} nome (label) da informação. Geralmente contém id da pessoa, tipo de pessoa, nome, sexo, funcionário público (S ou N) e código. Se for público, informa também esfera, órgão, cargo, uf.
\item \code{value} conteúdo da informação.
}
}
\description{
\code{\link[=cnc_parse_pag]{cnc_parse_pag()}} transforma arquivo HTML de uma página do CNC em um data.frame.

\code{\link[=cnc_parse_pessoa]{cnc_parse_pessoa()}} transforma HTMLs de pesquisas do tipo
\url{http://www.cnj.jus.br/improbidade_adm/visualizar_processo.php?seq_condenacao=1101}
em um data.frame com dados estruturados.

\code{\link[=cnc_parse_processo]{cnc_parse_processo()}} transforma HTMLs de pesquisas do tipo
\url{http://www.cnj.jus.br/improbidade_adm/visualizar_processo.php?seq_processo=9421}
em um data.frame com dados estruturados.

\code{\link[=cnc_parse_pessoa_infos]{cnc_parse_pessoa_infos()}} transforma HTMLs de pesquisas do tipo
\url{http://www.cnj.jus.br/improbidade_adm/visualizar_condenacao.php?seq_condenacao=1&rs=getDadosParte&rst=&rsrnd=0&rsargs[]=1}
em um data.frame com dados estruturados.
}
