% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-cnc.R
\name{parse_pags}
\alias{parse_pags}
\title{Página do CNC para dados.}
\usage{
parse_pags(arqs, verbose = TRUE)
}
\arguments{
\item{arqs}{character vector dos caminhos de arquivos a serem transformados.}

\item{verbose}{se \code{TRUE} (default), mostra o arquivo que está sendo lido com probabilidade de 5\%.}
}
\value{
\code{data.frame} com as colunas
\itemize{
  \item \code{arq} nome do arquivo lido.
  \item \code{id} id (1 a 15) do indivíduo obtido na página.
  \item \code{key} \code{nm_pessoa} ou \code{num_processo}, indicando se é o nome da pessoa ou número do processo.
  \item \code{value} nome da pessoa ou número do processo.
  \item \code{link} para acessar informações da pessoa ou do processo (info utilizada nos outros scrapers).
  \item \code{result} indica se o algoritmo funcionou ('OK') ou deu erro ('erro')
}
}
\description{
Transforma arquivo HTML de uma página do CNC em um data.frame.
}
