% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-cnc.R
\name{parse_pessoas}
\alias{parse_pessoas}
\title{Página de resultados de pessoas para dados.}
\usage{
parse_pessoas(arqs, verbose = TRUE)
}
\arguments{
\item{arqs}{character vector dos caminhos de arquivos a serem transformados.}

\item{verbose}{se \code{TRUE} (default), mostra o arquivo que está sendo lido com probabilidade de 5\%.}
}
\value{
\code{data_frame} com as colunas
\itemize{
  \item \code{arq} nome do arquivo.
  \item \code{key} nome (label) da informação. Geralmente contém infos da data de cadastro, dados do processo, dados da pessoa e dados da condenação.
  \item \code{value} conteúdo da informação.
  \item \code{result} indica se o algoritmo funcionou ('OK') ou deu erro ('erro')
}
}
\description{
Transforma HTMLs de pesquisas do tipo
\url{http://www.cnj.jus.br/improbidade_adm/visualizar_condenacao.php?seq_condenacao=1101}
em um data.frame com dados estruturados.
}
