% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-cnc.R
\name{parse_processos}
\alias{parse_processos}
\title{Página de resultados de processos para dados.}
\usage{
parse_processos(arqs, verbose = TRUE)
}
\arguments{
\item{arqs}{character vector dos caminhos de arquivos a serem transformados.}

\item{verbose}{se \code{TRUE} (default), mostra o arquivo que está sendo lido com probabilidade de 5\%.}
}
\value{
\code{data_frame} com as colunas
\itemize{
  \item \code{arq} nome do arquivo.
  \item \code{key} nome (label) da informação. Geralmente contém localização do processo, data da propositura e número do processo.
  \item \code{value} conteúdo da informação.
  \item \code{result} indica se o algoritmo funcionou ('OK') ou deu erro ('erro')
}
}
\description{
Transforma HTMLs de pesquisas do tipo
\url{http://www.cnj.jus.br/improbidade_adm/visualizar_processo.php?seq_processo=9421}
em um data.frame com dados estruturados.
}
