% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_pessoas}
\alias{tidy_pessoas}
\title{Tidyfica base que vem de parse_cnc_pessoas_infos}
\usage{
tidy_pessoas(cnc_pessoa_infos)
}
\arguments{
\item{cnc_pessoa_infos}{base raw cnc_pessoa_infos}
}
\description{
Tidyfica base que vem de parse_cnc_pessoas_infos
}
