% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-cnc.R
\name{tidy_processos}
\alias{tidy_processos}
\title{Tidyfica base que vem de parse_cnc_processos}
\usage{
tidy_processos(cnc_processos)
}
\arguments{
\item{cnc_processos}{base de dados raw dos processos.}
}
\description{
Tidyfica base que vem de parse_cnc_processos.
}

