% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-cnc.R
\name{tidy_cnc}
\alias{tidy_cnc}
\title{Tidyfica base que vem de todas as bases}
\format{A data frame with 35977 rows and 57 variables:
\describe{
  \item{a}{a}
  \item{b}{b}
  ...
}}
\source{
\url{https://www.cnj.jus.br/improbidade_adm/consultar_requerido.php}
}
\usage{
tidy_cnc(cnc_condenacoes, cnc_pags, cnc_processos, cnc_pessoa_infos)

tidy_cnc(cnc_condenacoes, cnc_pags, cnc_processos, cnc_pessoa_infos)
}
\description{
Tidyfica base que vem de todas as bases

Banco de dados contendo 57 infos de uma base de 35,977 variaveis de interesse.
}

